Move deps inside flags
